pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aayolo" 
version = "0.1.0"
description = "Anomaly-Aware YOLO for Infrared Small Target Detection" 
readme = "README.md"
requires-python = ">=3.9"
authors = [
  {name = "Alina Ciocarlan", email = "alina.ciocarlan@polytechnique.edu"},
]
dependencies = [
    "ipython>=7.34.0",
    "matplotlib>=3.2.2",
    "numpy>=1.18.5,<1.24.0",
    "opencv-python>=4.1.1",
    "pandas>=1.1.4",
    "pillow>=7.1.2",
    "protobuf<4.21.3",
    "psutil>=7.2.2",
    "pyyaml>=5.3.1",
    "requests>=2.23.0",
    "scipy>=1.4.1",
    "seaborn>=0.11.0",
    "tensorboard>=2.4.1",
    "thop>=0.1.1.post2209072238",
    "torch>=1.7.0,!=1.12.0",
    "torchvision>=0.8.1,!=0.13.0",
    "tqdm>=4.41.0",
]

[tool.setuptools.packages.find]
include = ["models*", "utils*"]

[tool.ruff]
target-version = "py39"
line-length = 120
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501", "E741"]

[tool.ruff.isort]
known-first-party = ["models", "utils"]

[tool.black]
line-length = 120
target-version = ["py39"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
